[WIP-HDR]: gpu: hdmi: Always install colorspace connector property
The colorspace property let user read/set the supported colorspaces which
should be filtered against the EDID data. Without this, the compositors
cannot know and signal its supported and what is going to be used.

FIXME: We also trigger an update on HDR data change, should be its own
patch

Signed-off-by: Nicolas Dufresne <nicolas.dufresne@collabora.com>