FROMLIST(v7): arm64: dts: rockchip: Use a longer PWM period for the fan on Radxa ROCK 5C
The fan on Radxa ROCK 5C is driven via an AO3416 MOSFET, which has a
total switch-on time of 0.6us and a total switch-off time of 6us [1],
meaning that the current PWM period of just 10us is too short for
fine-grained fan speed control. Increase the PWM period to 60us, so
that the switch-on and switch-off time of the MOSFET fall within a
more reasonable ~10% of the full period, thus making lower PWM duty
cycles meaningful.

[1] https://www.aosmd.com/pdfs/datasheet/AO3416.pdf

Fixes: 3ddf5cdb77e6 ("arm64: dts: rockchip: add Radxa ROCK 5C")
Reviewed-by: Dragan Simic <dsimic@manjaro.org>
Signed-off-by: FUKAUMI Naoki <naoki@radxa.com>
---
Changes in v7:
- Collect R-b tag
- Reword commit message
Changes in v6:
- None
Changes in v5:
- Reword commit message (Alexey Charkov)
Changes in v4:
- None
Changes in v3:
- None
Changes in v2:
- Reword commit message