LOCAL: arm64: dts: rockchip: rock5b: disable sdio node
The rock-5b board ships with an M2 slot that can be used with PCIe or SDIO
cards so both are described and enabled in device-tree. While there are a
huge number of inexpensive PCIe M2 cards available, there are few SDIO M2
cards in circulation, and none are available from Radxa.

Reflect real-world usage of the 5B board; leave the SDIO node described as
a capability but disable it by default to prevent dmesg errors seen during
mmc probe for an SDIO card that in-practice never exists.

ROCK5B:~ # dmesg | grep mmc2
[    0.790097] mmc_host mmc2: card is non-removable.
[    0.804379] mmc_host mmc2: Bus speed (slot 0) = 400000Hz (slot req 400000Hz, actual 400000HZ div = 0)
[    1.968538] mmc_host mmc2: Timeout sending command (cmd 0x202000 arg 0x0 status 0x80202000)
[    1.992757] mmc_host mmc2: Bus speed (slot 0) = 300000Hz (slot req 300000Hz, actual 300000HZ div = 0)
[    3.163937] mmc_host mmc2: Timeout sending command (cmd 0x202000 arg 0x0 status 0x80202000)
[    3.177872] mmc_host mmc2: Bus speed (slot 0) = 200000Hz (slot req 200000Hz, actual 200000HZ div = 0)
[    4.359405] mmc_host mmc2: Timeout sending command (cmd 0x202000 arg 0x0 status 0x80202000)
[    4.373304] mmc_host mmc2: Bus speed (slot 0) = 187500Hz (slot req 187500Hz, actual 187500HZ div = 0)
[    5.538223] mmc_host mmc2: Timeout sending command (cmd 0x202000 arg 0x0 status 0x80202000)
[    5.539621] mmc2: Failed to initialize a non-removable card

Signed-off-by: Christian Hewitt <christianshewitt@gmail.com>